mcp-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

python-dotenv==1.0.1
PyYAML==6.0.2

pydantic==2.10.5
requests==2.32.3

google-generativeai==0.8.3

Pillow==10.4.0

websockets

python-dateutil==2.9.0.post0

fpdf2